% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bariatric}
\alias{bariatric}
\title{Bariatric surgery data}
\format{
A list with three elements:
\describe{
  \item{X.pqn}{Matrix of pre-processed Proton NMR spectra with rows representing spectra}
  \item{ppm}{Chemical shift vector in ppm with its length equals the number of columns in \code{X.pqn}}
  \item{meta}{Spectrometer metadata}
}
}
\source{
Dr. Jia V. Li, Imperial College London, Department of Surgery and Cancer, Centre for Computational and Systems Medicine. \url{https://www.imperial.ac.uk/people/jia.li}
}
\usage{
data(bariatric)
}
\description{
Metabolic profiling NMR dataset of murine urine samples collected before and after bariatric surgery (Roux-en-Y gastric bypass).
}
\details{
One dimensional proton NMR spectra of murine urine samples, acquired on a Bruker Avance III spectrometer (Bruker, Rheinstetten, Germany) operating at a frequency of 600.13 MHz and a temperature of 300 K. A standard NMR pulse sequence (\emph{recycle delay-90°-t1-90°-tm-90°-acquisition}) was applied to acquire one-dimensional Proton NMR spectral data, where \emph{t1} was set to 3 μs and \emph{tm} (mixing time) was set to 100 ms. Water suppression was achieved using selective irradiation during a \emph{recycle delay} of 2 s and \emph{tm}. A 90° pulse was adjusted to 10 μs. A total of 128 scans were collected into 64 k data points with a spectral width of 20 ppm.

For further information on sample collection and processing please see reference below. Pre-processing of raw NMR data is described in the \emph{MetaboMate} vignette Data Import and Preprocessing.
}
\references{
Li, Jia V., \emph{et al.} (2011) Metabolic surgery profoundly influences gut microbial-host metabolic cross-talk. \emph{Gut}. 60.9, 1214-1223.

\url{https://www.ncbi.nlm.nih.gov/pubmed/21572120}
}
\author{
Torben Kimhofer \email{tkimhofer@murdoch.edu.au}
}
\keyword{datasets}
